requirements.txt
djangorestframework=3.11.0
django = 3.0.8
djangorestframework-xml